transport/test_talk/producer/requirements.txt
tinvest
kafka-python